+++ controllers/auth/index.js
@@ -1,9 +1,11 @@
 const register = require("./register");
 const login = require("./login");
 const logout = require("./logout");
+const getUserInfo = require("./getUserInfo");
 
 module.exports = {
   register,
   login,
   logout,
+  getUserInfo,
 };
