+++ controllers/pets/addPet.js
@@ -3,15 +3,9 @@ const { Pets } = require("../../models/pets");
 // тут не req.body а прийом з форми
 
 const addPet = async (req, res) => {
-  try {
-    const { _id: owner } = req.user;
-    const addNewPet = await Pets.create({ ...req.body, owner });
-    res.status(201).json(addNewPet);
-  } catch (error) {
-    res.status(500).json({
-      message: error.message,
-    });
-  }
+  const { _id: owner } = req.user;
+  const addNewPet = await Pets.create({ ...req.body, owner });
+  res.status(201).json(addNewPet);
 };
 
 module.exports = addPet;
