+++ controllers/pets/index.js
@@ -1,0 +1,9 @@
+const addPet = require("./addPet");
+const getUserPets = require("./getUserPets");
+const removePet = require("./removePet");
+
+module.exports = {
+  addPet,
+  getUserPets,
+  removePet,
+};
