+++ routes/api/users.js
@@ -1,8 +1,9 @@
 const express = require("express");
 const getUserInfo = require("../../controllers/auth/getUserInfo.js");
+const authenticate = require("../../middlewares/authenticate");
 
 const router = express.Router();
 
-router.get("/current", getUserInfo);
+router.get("/current", authenticate, getUserInfo);
 
 module.exports = router;
